# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/stuartcrobinson/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/indexes/{indexName}/synonyms/batch": {
      "post": {
        "tags": [
          "synonyms"
        ],
        "summary": "Batch create or update multiple synonym entries in the specified index.",
        "operationId": "save_synonyms",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Array of synonyms",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Synonyms saved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid query parameter value"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/indexes/{indexName}/synonyms/clear": {
      "post": {
        "tags": [
          "synonyms"
        ],
        "summary": "Remove all synonym entries from the specified index.",
        "operationId": "clear_synonyms",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Synonyms cleared",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/indexes/{indexName}/synonyms/search": {
      "post": {
        "tags": [
          "synonyms"
        ],
        "summary": "Search for synonym entries in the specified index with optional filtering.",
        "operationId": "search_synonyms",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Search parameters",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Matching synonyms",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/indexes/{indexName}/synonyms/{objectID}": {
      "get": {
        "tags": [
          "synonyms"
        ],
        "summary": "Retrieve a synonym entry by its object ID from the specified index.",
        "operationId": "get_synonym",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objectID",
            "in": "path",
            "description": "Synonym ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Synonym retrieved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Synonym not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "synonyms"
        ],
        "summary": "Save or update a single synonym entry in the specified index.",
        "operationId": "save_synonym",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objectID",
            "in": "path",
            "description": "Synonym ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Synonym data",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Synonym saved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "synonyms"
        ],
        "summary": "Delete a synonym entry from the specified index by its object ID.",
        "operationId": "delete_synonym",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objectID",
            "in": "path",
            "description": "Synonym ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Synonym deleted",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Synonym not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}